third_party/WebKit/LayoutTests/http/tests/preload/document_write_preload.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
var t = async_test('Simple scripts that inject external scripts via document.write should be preloaded');
// We reject scripts with "for"
window.perf = window.performance;
</script>
<script>
if (window.perf)
    var boundedStart = window.perf.now();
var src = '../resources/dummy.js';
document.write('<scr' + 'ipt src="' + src + '"></scr' + 'ipt>');
</script>
<script>
window.addEventListener("load", t.step_func(function() {
    var resourceTiming = window.performance.getEntriesByType('resource')[2];
    assert_less_than(resourceTiming.startTime, boundedStart);
    t.done();
}));
</script>
